refactor:custom header component created
